Avoid a couple of unneeded ixs to reduce the tx size
I don't understand why the Approve and Revoke instructions were supposedly needed to charge the payer via a delegate disposable wallet instead of charging directly to the payer when using an SPL Token as the payment method.

Removing these two instructions means the minting and cleanup tx sizes are reduced and, depending on the CM config, increases the chances of being able to mint in a single tx.

This is the case of a config like the following:

```json
{
  "price": 100,
  "number": 10,
  "gatekeeper": null,
  "solTreasuryAccount": null,
  "splTokenAccount": "<<<<<<--- SOME ACCOUNT HERE --->>>>>>",
  "splToken": "<<<<<<--- SOME TOKEN HERE --->>>>>>",
  "goLiveDate": "30 Apr 2022 00:00:00 GMT",
  "endSettings": null,
  "whitelistMintSettings": null,
  "hiddenSettings": null,
  "storage": "arweave",
  "ipfsInfuraProjectId": null,
  "ipfsInfuraSecret": null,
  "awsS3Bucket": null,
  "nftStorageKey": null,
  "noRetainAuthority": false,
  "noMutable": false
}
```

Given a CM using a config like that, plus using a collection so that the collection is set during mint, 3 txs were needed before this commit (setup, mint, and cleanup), but only one is needed with this change.

Here's an NFT minted with this change, in a single tx:
https://solscan.io/token/DVBqW3CTTgLew5KJdiP4gB7JEpop7i1C1ZLyZ8m9QUz?cluster=devnet#txs